+++ includes/Abilities/CustomPostTypes.php
@@ -47,13 +47,7 @@ class CustomPostTypes {
 					$root = RestApiUtils::get_latest_available_rest_route( $this->base_namespace, 'types' );
 
 					if ( ! $root ) {
-						return blu_standardize_rest_response(
-							new \WP_Error(
-								400,
-								'A valid route for types not found. Please ensure that the REST API is enabled and that the latest version of the WordPress REST API is installed.',
-							)
-						);
-
+						return blu_standardize_route_unavailable_for_resource( 'types' );
 					}
 
 					$method     = 'GET';