+++ includes/Abilities/Media.php
@@ -346,12 +346,7 @@ class Media {
 		$root = RestApiUtils::get_latest_available_rest_route( $this->base_namespace, 'media' );
 
 		if ( ! $root ) {
-			return blu_standardize_rest_response(
-				new \WP_Error(
-					400,
-					'A valid route for media not found. Please ensure that the REST API is enabled and that the latest version of the WordPress REST API is installed.',
-				)
-			);
+			return blu_standardize_route_unavailable_for_resource( 'media' );
 		}
 
 		return $root;